refactor(client): home servvice
