+++ client/src/pages/home/home.service.tsx
@@ -1,9 +1,34 @@
 import { useMutation } from '@apollo/client';
+import { SIGN_IN_PATH } from 'constants/paths';
+import AuthContext from 'context/profile/auth.context';
 import { SignOutMutation, SignOutMutationVariables } from 'graphql/__generated__/graphql';
+import { getGraphQLErrorMessage } from 'helpers/get-graphql-error-message';
 import { SIGN_OUT_MUTATION } from 'pages/home/home.graphql';
+import { useContext } from 'react';
+import { useNavigate } from 'react-router-dom';
+import { toast } from 'react-toastify';
 
-export const useSignOut = (onSignOutSuccess: () => Promise<void>) => {
-	const [signOut, { loading }] = useMutation<SignOutMutation, SignOutMutationVariables>(
+const useOnSignOutSuccess = () => {
+	const navigate = useNavigate();
+
+	const { refetch: refetchProfile } = useContext(AuthContext);
+
+	const onSignOutSuccess = async () => {
+		try {
+			await refetchProfile();
+		} finally {
+			navigate(SIGN_IN_PATH);
+			toast.success(`You successfully signed out`);
+		}
+	};
+
+	return { onSignOutSuccess };
+};
+
+export const useSignOut = () => {
+	const { onSignOutSuccess } = useOnSignOutSuccess();
+
+	const [signOutMutation, { loading }] = useMutation<SignOutMutation, SignOutMutationVariables>(
 		SIGN_OUT_MUTATION,
 		{
 			onCompleted: async () => {
@@ -11,6 +36,15 @@ export const useSignOut = (onSignOutSuccess: () => Promise<void>) => {
 			}
 		}
 	);
+
+	const signOut = async () => {
+		try {
+			await signOutMutation();
+		} catch (error) {
+			toast.error(getGraphQLErrorMessage(error), { autoClose: 10000 });
+		}
+		return;
+	};
 
 	return { loading, signOut };
 };
